+++ clarify-app/package.json
@@ -11,8 +11,11 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "bootstrap": "^5.3.2",
+    "dotenv": "^16.4.4",
+    "openai": "^4.28.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-google-button": "^0.7.2",
     "react-router-dom": "^6.22.0",
     "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.4"
