+++ clarify-app/src/components/Navbar.jsx
@@ -1,6 +1,8 @@
 import React from 'react';
+import profileImage from "../assets/Corey.png";
 import { Link } from 'react-router-dom';
 import { UserAuth } from '../context/AuthContext';
+import "./NavBar.css";
 
 const Navbar = () => {
    const { user, logOut } = UserAuth();
@@ -14,17 +16,66 @@ const Navbar = () => {
    }
 
    return (
-      <div className='flex justify-between bg-gray-200 w-full p-4'>
-         <Link to='/'>
-            <h1 className='text-center text-2xl font-bold'>
-               Clarify
-            </h1>
-         </Link>
-         {user?.displayName ? (
-            <button onClick={handleSignOut}>Logout</button>
-         ) : (
-            <Link to='/signin'>Sign in</Link>
-         )}
+      <div className="NavBar_Container">
+         <nav className="navbar navbar-expand-lg bg-body-tertiary">
+            <div className="container-fluid">
+
+               <Link
+                  to="/"
+                  className="navbar-brand navbar-header">
+                  Clarify
+               </Link>
+
+               <button
+                  className="navbar-toggler dropdown-button"
+                  type="button"
+                  data-bs-toggle="collapse"
+                  data-bs-target="#navbarSupportedContent"
+                  aria-controls="navbarSupportedContent"
+                  aria-expanded="false"
+                  aria-label="Toggle navigation"
+               >
+                  <span className="navbar-toggler-icon"></span>
+               </button>
+               <div className="collapse navbar-collapse nav-dropdown" id="navbarSupportedContent">
+                  <ul className="navbar-nav me-auto mb-2 mb-lg-0">
+                     <li className="nav-item">
+                        <Link
+                           to="/"
+                           className="nav-link active" aria-current="page">
+                           Home
+                        </Link>
+                     </li>
+                     <li className="nav-item">
+                        <Link
+                           to="/account"
+                           className="nav-link active" aria-current="page">
+                           Account
+                        </Link>
+                     </li>
+                     <li className="nav-item">
+                        <Link
+                           to="/chat"
+                           className="nav-link active" aria-current="page">
+                           Chat
+                        </Link>
+                     </li>
+                     <li className="nav-item">
+                        {user?.displayName ? (
+                           <button className="nav-link active" onClick={handleSignOut}>Logout</button>
+                        ) : (
+                           <Link to='/signin'>Sign in</Link>
+                        )}
+                     </li>
+                  </ul>
+               </div>
+               {user?.displayName ? (
+                           <img className="profileImg" src={profileImage} alt="Profile" />
+                        ) : (
+                           ""
+                        )}
+            </div>
+         </nav>
       </div>
    );
 };
