+++ clarify-app/src/pages/Account.jsx
@@ -14,6 +14,13 @@ const Account = () => {
 
   return (
     <div className='w-[300px] m-auto'>
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
       <h1 className='text-center text-2xl font-bold pt-12'>Account</h1>
       <div>
         <p>Welcome, {user?.displayName}</p>
