+++ clarify-app/src/pages/LandingPage.css
@@ -9,7 +9,7 @@
 {
     width: 100%;
     height: 100vh;
-    background: url(../src/assets/bg-image2.jpeg);
+    background: url(../../src/assets/bg-image2.jpeg);
     background-size: cover;
     background-attachment: scroll;
     background-repeat: no-repeat;
